refactor(executor): split data-parallel into its own leaf module (SoC)
The ThreadScheduler core.rs mixed four concerns in 554 lines. The indexed
data-parallel fan-out (for_each_indexed, map_reduce_indexed, ~190 lines) is
a distinct responsibility from task submission/scope/lifecycle and now
corresponds exactly to the DataParallel seam role, so it moves to a
dedicated runtime/scheduler/data_parallel.rs leaf module.

Pure reorganization, no behavior change: the methods, their bounds, and
bodies are unchanged (verified by the unchanged indexed_* tests). core.rs
drops to 357 lines (under the 500 target) and now owns only core
scheduling; the benchmark runtime-source alias gains the new leaf so source
contracts still see the full runtime scheduler surface.
