+++ moirai-executor/src/schedule/runtime/scheduler/mod.rs
@@ -1,6 +1,7 @@
 //! Unified thread scheduler module.
 
 pub mod core;
+pub mod data_parallel;
 pub mod scope;
 
 #[cfg(feature = "scheduler-diagnostics")]
